Move View.kt to core-android module
