+++ feature-users/build.gradle
@@ -15,13 +15,10 @@ android {
 }
 
 dependencies {
-  implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-
-  kapt 'com.google.dagger:dagger-compiler:2.21'
-
   api project(':core')
   api project(':core-android')
 
+  kapt 'com.google.dagger:dagger-compiler:2.21'
   api 'com.google.dagger:dagger:2.21'
 
   api 'com.squareup.retrofit2:retrofit:2.5.0'
